requirements.txt
asciimatics==1.15.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
idna==3.10
oauthlib==3.2.2
pillow==11.0.0
pyfiglet==1.0.2
python-dateutil==2.9.0.post0
python3-discogs-client==2.7.1
requests==2.32.3
six==1.17.0
urllib3==2.3.0
wcwidth==0.2.13
